requirements/base.txt
Django==3.0.7
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
gunicorn==20.0.4
python-dotenv==0.13.0